refactor(sentry): split the selector's CLI and run-budget layers back under the cap
The fail-closed wrapping pushed scripts/sentry-autofix-select.mjs to 828 lines
against a 600-line soft cap. An earlier round in this PR had already split it —
to 598, two lines under — and the very next change went straight back through.
Sizing a split to barely clear the line is why it crossed twice, so this one
targets real headroom: 500, with 100 lines to spare.

Two seams, both leaves that import nothing from the selector, so the re-exports
that keep every public name on sentry-autofix-select.mjs cannot form an ESM
cycle:

- sentry-autofix-select-cli.mjs (224) — the boundary between the run and the
  workflow around it: the option contract, the help text, the report files the
  tracker reads back, and `--emit-verdict`. Nothing here decides whether a stub
  is selectable.
- sentry-autofix-select-instrument.mjs (248) — what BOUNDS and OBSERVES a run:
  the per-run read cap and its no-op guard, the gh counter, the throttle latch,
  the zero-state record shapes, and the DEGRADED and summary lines. The cap and
  the counter are one subject from two sides — the cap is the predicted spend
  the cost bound is built from, ghCalls is the measured one — and it puts the
  first pass's cap/guard pair in the same shape as the second look's, which
  already lives outside the selector.

`main` stays in the selector: it is the one function needing both halves, and
the workflow invokes that file directly at three call sites, so the executable
entry point must not move. Verified by running it, not reading it — help, an
unknown option, a batch run against a stub gh, a throttled run, a hard list
failure and `--emit-verdict` all produce byte-identical output and exit codes
before and after the split.

Also routes the four modules this PR added through the quality gate.
sentry-autofix-second-look.mjs and sentry-autofix-decisions.mjs matched no case
at all, so editing either alone ran no suite; the two new modules join them, and
the two that own per-run cost caps also run the finalize suite, whose
timeout-minutes pin derives the budget from those live constants.
